chore: rename command to align with vault product
